move uuids out
